Refactor Animation for heading and create a custom hook
