+++ src/components/work2.js
@@ -5,8 +5,9 @@ import rollingPlansLogo from "../../public/images/rollingplans.png";
 import Image from "next/image";
 import WorkSocial from "./WorkSocial";
 import Work2Card from "./Work2Card";
+import useOnScreen from "../customHooks/useOnScreen";
 
-function work2({ works, useOnScreen }) {
+function work2({ works }) {
   const [ref, visible] = useOnScreen({ rootMargin: "-10px" });
   return (
     <section ref={ref} className='pb-20 lg:px-48 bg-personal_blue px-4'>
